kubeflow_components/predict-metadata-op/requirements-gpu.txt
opencv-python==4.5.3.56
decord
lxml
tqdm
absl-py
matplotlib
easydict
pillow
